remove dependency on Fabric API
