+++ platforms/fabric/src/main/resources/fabric.mod.json
@@ -25,7 +25,6 @@
   ],
   "depends": {
     "fabricloader": ">=0.7.4",
-    "fabric": "*",
     "minecraft": "1.16.x"
   },
   "accessWidener": "terra.accesswidener"
